Module Tezos_brassaia_context

module Brassaia_context : sig ... end

Brassaia_context is a persisting recursive map backed by an on-disk Brassaia tree, i.e. Tezos_context.Context (in src/lib_context), used by nodes. It is the "main" kind of context. It abstracts away Tezos_context.Context, by hiding the internals, and making it an instance of the generic Protocol_environment.Context.t. Other *_context modules of Tezos_protocol_environment, i.e. siblings of this file, are backed by different type of values coming from Tezos_context.