Module Tezos_client_base.Client_confirmations

val wait_for_operation_inclusion : Client_context.full -> chain:Tezos_shell_services.Chain_services.chain -> ?predecessors:int -> ?confirmations:int -> ?branch:Tezos_base.TzPervasives.Block_hash.t -> Tezos_base.TzPervasives.Operation_hash.t -> (Tezos_base.TzPervasives.Block_hash.t * int * int) Tezos_base.TzPervasives.tzresult Lwt.t

wait_for_operation_inclusion chain ~predecessors ~confirmations oph waits for `oph` to appears in the main chain with at least `confirmations`. It returns the hash of the block that contains the operation and the operation position in the block.

This functions also looks for the operations in the `predecessors` of the initial chain head.

val lookup_operation_in_previous_blocks : Client_context.full -> chain:Tezos_shell_services.Block_services.chain -> predecessors:int -> Tezos_base.TzPervasives.Operation_list_hash.elt -> (Tezos_base.TzPervasives.Block_hash.t * int * int) option Tezos_base.TzPervasives.tzresult Lwt.t

lookup an operation in predecessors previous blocks, starting from head

val wait_for_bootstrapped : ?retry: ((Client_context.full as 'a -> ((Tezos_base.TzPervasives.Block_hash.t * Tezos_base.TzPervasives.Time.Protocol.t) Lwt_stream.t * Tezos_rpc.Context.stopper) Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a -> ((Tezos_base.TzPervasives.Block_hash.t * Tezos_base.TzPervasives.Time.Protocol.t) Lwt_stream.t * Tezos_rpc.Context.stopper) Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t

returns when the node consider itself as bootstrapped.

Function retry specifies how to behave in order to connect to the node. The default is the identity which correspond to simply calling the RPC. As an example, the baker tries 5 times with delays in between attempts when the connection fails.