Module Client_keys.Signature_type

Signature_type is a small module to be included in signer to conform to the module type SIGNER instead of rewriting all type.

type nonrec public_key_hash = Tezos_crypto.Signature.V_latest.Public_key_hash.t

type nonrec public_key = Tezos_crypto.Signature.V_latest.Public_key.t

type nonrec secret_key = Tezos_crypto.Signature.V_latest.Secret_key.t

type nonrec signature = Tezos_crypto.Signature.V_latest.t

type nonrec pk_uri = pk_uri

type nonrec sk_uri = sk_uri