Module Client_keys.V1

module Signature_type : sig ... end

Signature_type is a small module to be included in signer to conform to the module type SIGNER instead of rewriting all type.

module Public_key_hash : Client_aliases.Alias with type t = Tezos_crypto.Signature.V1.Public_key_hash.t

module Public_key : Client_aliases.Alias with type t = pk_uri * Tezos_crypto.Signature.V1.Public_key.t option

module Secret_key : Client_aliases.Alias with type t = sk_uri

val import_secret_key : io:Client_context.io_wallet -> pk_uri -> (Tezos_crypto.Signature.V1.Public_key_hash.t * Tezos_crypto.Signature.V1.Public_key.t option) Tezos_base.TzPervasives.tzresult Lwt.t

val public_key : pk_uri -> Tezos_crypto.Signature.V1.Public_key.t Tezos_base.TzPervasives.tzresult Lwt.t

val public_key_hash : pk_uri -> (Tezos_crypto.Signature.V1.Public_key_hash.t * Tezos_crypto.Signature.V1.Public_key.t option) Tezos_base.TzPervasives.tzresult Lwt.t

val neuterize : sk_uri -> pk_uri Tezos_base.TzPervasives.tzresult Lwt.t

val sign : Client_context.wallet -> ?watermark:Tezos_crypto.Signature.V1.watermark -> sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_crypto.Signature.V1.t Tezos_base.TzPervasives.tzresult Lwt.t

val append : Client_context.wallet -> ?watermark:Tezos_crypto.Signature.V1.watermark -> sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Bytes.t Tezos_base.TzPervasives.tzresult Lwt.t

val check : ?watermark:Tezos_crypto.Signature.V1.watermark -> pk_uri -> Tezos_crypto.Signature.V1.t -> Tezos_base.TzPervasives.Bytes.t -> bool Tezos_base.TzPervasives.tzresult Lwt.t

val deterministic_nonce : sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Bytes.t Tezos_base.TzPervasives.tzresult Lwt.t

val deterministic_nonce_hash : sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Bytes.t Tezos_base.TzPervasives.tzresult Lwt.t

val supports_deterministic_nonces : sk_uri -> bool Tezos_base.TzPervasives.tzresult Lwt.t

val register_key : Client_context.wallet -> ?force:bool -> (Tezos_crypto.Signature.V1.Public_key_hash.t * pk_uri * sk_uri) -> ?public_key:Tezos_crypto.Signature.V1.Public_key.t -> string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val register_keys : Client_context.wallet -> (string * Tezos_crypto.Signature.V1.Public_key_hash.t * Tezos_crypto.Signature.V1.Public_key.t * pk_uri * sk_uri) list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Similar to repeated calls to register_key, but is more efficient. Always forces addition of new elements.

val list_keys : Client_context.wallet -> (string * Tezos_crypto.Signature.V1.Public_key_hash.t * Tezos_crypto.Signature.V1.Public_key.t option * sk_uri option) list Tezos_base.TzPervasives.tzresult Lwt.t

val alias_keys : Client_context.wallet -> string -> (Tezos_crypto.Signature.V1.Public_key_hash.t * Tezos_crypto.Signature.V1.Public_key.t option * sk_uri option) option Tezos_base.TzPervasives.tzresult Lwt.t

val get_key : Client_context.wallet -> Tezos_crypto.Signature.V1.Public_key_hash.t -> (string * Tezos_crypto.Signature.V1.Public_key.t * sk_uri) Tezos_base.TzPervasives.tzresult Lwt.t

val get_public_key : Client_context.wallet -> Tezos_crypto.Signature.V1.Public_key_hash.t -> (string * Tezos_crypto.Signature.V1.Public_key.t) Tezos_base.TzPervasives.tzresult Lwt.t

val get_keys : Client_context.wallet -> (string * Tezos_crypto.Signature.V1.Public_key_hash.t * Tezos_crypto.Signature.V1.Public_key.t * sk_uri) list Tezos_base.TzPervasives.tzresult Lwt.t

val force_switch : unit -> (bool, 'ctx) Tezos_clic.arg