Module type Client_keys.COMMON_SIGNER

type pk_uri = private Uri.t

type sk_uri = private Uri.t

type public_key_hash

type public_key

type secret_key

type signature

val scheme : string

scheme is the name of the scheme implemented by this signer module.

val title : string

title is a one-line human readable description of the signer.

val description : string

description is a multi-line human readable description of the signer, that should include the format of key specifications.

val neuterize : sk_uri -> pk_uri Tezos_base.TzPervasives.tzresult Lwt.t

neuterize sk is the corresponding pk.

val import_secret_key : io:Client_context.io_wallet -> pk_uri -> (public_key_hash * public_key option) Tezos_base.TzPervasives.tzresult Lwt.t

import_secret_key ~io pk is the function to be called when interactively importing a key-pair and returning the public key and its hash.

Some signer implementations improve long-term security by requiring human/manual validation while importing keys, the ~io argument can be used to prompt the user in such case.

val public_key : pk_uri -> public_key Tezos_base.TzPervasives.tzresult Lwt.t

public_key pk is the Ed25519 version of pk.

val public_key_hash : pk_uri -> (public_key_hash * public_key option) Tezos_base.TzPervasives.tzresult Lwt.t

public_key_hash pk is the hash of pk. As some signers will query the full public key to obtain the hash, it can be optionally returned to reduce the amount of queries.