Module Tezos_client_base.Client_keys_v0

Module to use keys over Tezos_crypto.Signature.V0

include module type of struct include Client_keys end

Cryptographic keys tables

type pk_uri = private Uri.t

module Pk_uri_hashtbl = Client_keys.Pk_uri_hashtbl

type sk_uri = private Uri.t

type sapling_uri = private Uri.t

val pk_uri_parameter : unit -> (pk_uri, 'a) Tezos_clic.parameter

val pk_uri_param : ?name:string -> ?desc:string -> ('a, 'b) Tezos_clic.params -> (pk_uri -> 'a, 'b) Tezos_clic.params

val sk_uri_parameter : unit -> (sk_uri, 'a) Tezos_clic.parameter

val sk_uri_param : ?name:string -> ?desc:string -> ('a, 'b) Tezos_clic.params -> (sk_uri -> 'a, 'b) Tezos_clic.params

type Tezos_base.TzPervasives.error +=

| Unregistered_key_scheme of string

type Tezos_base.TzPervasives.error +=

| Invalid_uri of Uri.t

type sapling_key = Client_keys.sapling_key = {

sk : sapling_uri;

path : int32 list;

address_index : Tezos_sapling.Core.Client.Viewing_key.index;

}

module Sapling_key = Client_keys.Sapling_key

module Logging = Client_keys.Logging

Interface for external signing modules.

module type COMMON_SIGNER = Client_keys.COMMON_SIGNER

module type SIGNER = Client_keys.SIGNER

type signer = (module SIGNER)

val register_signer : (module SIGNER) -> unit

register_signer signer registers first-class module signer as signer for keys with scheme (val signer : SIGNER).scheme.

val registered_signers : unit -> (string * signer) list

module type S = Client_keys.S

module V0 = Client_keys.V0

module V1 = Client_keys.V1

module V_latest = Client_keys.V_latest

include module type of V_latest

val string_of_sk_uri : sk_uri -> string

val sk_uri_of_string : string -> sk_uri

include module type of struct include Client_keys.V0 end

module Signature_type = Client_keys.V0.Signature_type

Signature_type is a small module to be included in signer to conform to the module type SIGNER instead of rewriting all type.

module Public_key_hash = Client_keys.V0.Public_key_hash

module Public_key = Client_keys.V0.Public_key

module Secret_key = Client_keys.V0.Secret_key

val import_secret_key : io:Client_context.io_wallet -> Client_keys.pk_uri -> (Tezos_crypto.Signature.V0.Public_key_hash.t * Tezos_crypto.Signature.V0.Public_key.t option) Tezos_base.TzPervasives.tzresult Lwt.t

val public_key : Client_keys.pk_uri -> Tezos_crypto.Signature.V0.Public_key.t Tezos_base.TzPervasives.tzresult Lwt.t

val public_key_hash : Client_keys.pk_uri -> (Tezos_crypto.Signature.V0.Public_key_hash.t * Tezos_crypto.Signature.V0.Public_key.t option) Tezos_base.TzPervasives.tzresult Lwt.t

val neuterize : Client_keys.sk_uri -> Client_keys.pk_uri Tezos_base.TzPervasives.tzresult Lwt.t

val sign : Client_context.wallet -> ?watermark:Tezos_crypto.Signature.V0.watermark -> Client_keys.sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_crypto.Signature.V0.t Tezos_base.TzPervasives.tzresult Lwt.t

val append : Client_context.wallet -> ?watermark:Tezos_crypto.Signature.V0.watermark -> Client_keys.sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Bytes.t Tezos_base.TzPervasives.tzresult Lwt.t

val check : ?watermark:Tezos_crypto.Signature.V0.watermark -> Client_keys.pk_uri -> Tezos_crypto.Signature.V0.t -> Tezos_base.TzPervasives.Bytes.t -> bool Tezos_base.TzPervasives.tzresult Lwt.t

val deterministic_nonce : Client_keys.sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Bytes.t Tezos_base.TzPervasives.tzresult Lwt.t

val deterministic_nonce_hash : Client_keys.sk_uri -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Bytes.t Tezos_base.TzPervasives.tzresult Lwt.t

val supports_deterministic_nonces : Client_keys.sk_uri -> bool Tezos_base.TzPervasives.tzresult Lwt.t

val register_key : Client_context.wallet -> ?force:bool -> (Tezos_crypto.Signature.V0.Public_key_hash.t * Client_keys.pk_uri * Client_keys.sk_uri) -> ?public_key:Tezos_crypto.Signature.V0.Public_key.t -> string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val register_keys : Client_context.wallet -> (string * Tezos_crypto.Signature.V0.Public_key_hash.t * Tezos_crypto.Signature.V0.Public_key.t * Client_keys.pk_uri * Client_keys.sk_uri) list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Similar to repeated calls to register_key, but is more efficient. Always forces addition of new elements.

val list_keys : Client_context.wallet -> (string * Tezos_crypto.Signature.V0.Public_key_hash.t * Tezos_crypto.Signature.V0.Public_key.t option * Client_keys.sk_uri option) list Tezos_base.TzPervasives.tzresult Lwt.t

val alias_keys : Client_context.wallet -> string -> (Tezos_crypto.Signature.V0.Public_key_hash.t * Tezos_crypto.Signature.V0.Public_key.t option * Client_keys.sk_uri option) option Tezos_base.TzPervasives.tzresult Lwt.t

val get_key : Client_context.wallet -> Tezos_crypto.Signature.V0.Public_key_hash.t -> (string * Tezos_crypto.Signature.V0.Public_key.t * Client_keys.sk_uri) Tezos_base.TzPervasives.tzresult Lwt.t

val get_public_key : Client_context.wallet -> Tezos_crypto.Signature.V0.Public_key_hash.t -> (string * Tezos_crypto.Signature.V0.Public_key.t) Tezos_base.TzPervasives.tzresult Lwt.t

val get_keys : Client_context.wallet -> (string * Tezos_crypto.Signature.V0.Public_key_hash.t * Tezos_crypto.Signature.V0.Public_key.t * Client_keys.sk_uri) list Tezos_base.TzPervasives.tzresult Lwt.t

val force_switch : unit -> (bool, 'ctx) Tezos_clic.arg