Module Smart_rollup_alias.Address

include Client_aliases.Alias with type t = Tezos_crypto.Hashed.Smart_rollup_address.t

type t = Tezos_crypto.Hashed.Smart_rollup_address.t

type fresh_param

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val load : Client_context.wallet -> (string * t) list Tezos_base.TzPervasives.tzresult Lwt.t

val set : Client_context.wallet -> (string * t) list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val find : Client_context.wallet -> string -> t Tezos_base.TzPervasives.tzresult Lwt.t

val find_opt : Client_context.wallet -> string -> t option Tezos_base.TzPervasives.tzresult Lwt.t

val rev_find : Client_context.wallet -> t -> string option Tezos_base.TzPervasives.tzresult Lwt.t

val rev_find_all : Client_context.wallet -> t -> string list Tezos_base.TzPervasives.tzresult Lwt.t

val name : Client_context.wallet -> t -> string Tezos_base.TzPervasives.tzresult Lwt.t

val mem : Client_context.wallet -> string -> bool Tezos_base.TzPervasives.tzresult Lwt.t

val add : force:bool -> Client_context.wallet -> string -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val add_many : Client_context.wallet -> (string * t) list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Similar to repeated calls to add, but more efficient. Always forces addition of new elements.

val del : Client_context.wallet -> string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val update : Client_context.wallet -> string -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val of_source : string -> t Tezos_base.TzPervasives.tzresult Lwt.t

val to_source : t -> string Tezos_base.TzPervasives.tzresult Lwt.t

val alias_parameter : unit -> (string * t, Client_context.wallet) Tezos_clic.parameter

val alias_param : ?name:string -> ?desc:string -> ('a, Client_context.wallet as 'b) Tezos_clic.params -> ((string * t) -> 'a, 'b) Tezos_clic.params

val aliases_param : ?name:string -> ?desc:string -> ('a, Client_context.wallet as 'b) Tezos_clic.params -> ((string * t) list -> 'a, 'b) Tezos_clic.params

val fresh_alias_param : ?name:string -> ?desc:string -> ('a, < .. > as 'obj) Tezos_clic.params -> (fresh_param -> 'a, 'obj) Tezos_clic.params

val force_switch : unit -> (bool, _) Tezos_clic.arg

val of_fresh : Client_context.wallet -> bool -> fresh_param -> string Tezos_base.TzPervasives.tzresult Lwt.t

val parse_source_string : Client_context.wallet -> string -> t Tezos_base.TzPervasives.tzresult Lwt.t

val source_param : ?name:string -> ?desc:string -> ('a, Client_context.wallet as 'obj) Tezos_clic.params -> (t -> 'a, 'obj) Tezos_clic.params

val source_arg : ?long:string -> ?placeholder:string -> ?doc:string -> unit -> (t option, Client_context.wallet as 'obj) Tezos_clic.arg

val autocomplete : Client_context.wallet -> string list Tezos_base.TzPervasives.tzresult Lwt.t

module Entity : Client_aliases.Entity with type t = Tezos_crypto.Hashed.Smart_rollup_address.t

val parameter : unit -> (t, Client_context.wallet) Tezos_clic.parameter

Argument parser for rollup address literals or aliases.

val param : ?name:string -> ?desc:string -> ('a, Client_context.wallet as 'obj) Tezos_clic.params -> (t -> 'a, 'obj) Tezos_clic.params

Command line parameter for rollup address literals or aliases.