Module Tezos_duo_context_lib

module Context_wrapper : sig ... end

module Duo_context : sig ... end

Brassaia_context is a persisting recursive map backed by an on-disk Brassaia tree, i.e. Tezos_context.Context (in src/lib_context), used by nodes. It is the "main" kind of context. It abstracts away Tezos_context.Context, by hiding the internals, and making it an instance of the generic Protocol_environment.Context.t. Other *_context modules of Tezos_protocol_environment, i.e. siblings of this file, are backed by different type of values coming from Tezos_context.

module Duo_context_lib : sig ... end

module Duo_memory_context : sig ... end

Memory_context is a recursive map backed by an in-memory Brassaia tree, i.e. Tezos_context_memory.Context (in src/lib_context), used by the mockup mode and by proof verifiers (i.e. clients of Tezos_context_helpers.Context.Make_tree.Proof). It abstracts away Tezos_context_memory.Context.t.