Module Tezos_mockup.Local_services

module Directory : sig ... end

type Tezos_base.TzPervasives.error +=

| Injection_not_possible

type Tezos_base.TzPervasives.error +=

| Cannot_parse_op

type Tezos_base.TzPervasives.error +=

| Cannot_parse_proto_data

type callback_writer = Tezos_protocol_environment.rpc_context -> bytes -> unit Tezos_base.TzPervasives.tzresult Lwt.t

module type MENV = sig ... end

module Make (E : MENV) : sig ... end

val build_shell_directory : string -> Tezos_mockup_registration.Registration.mockup_environment -> Tezos_base.TzPervasives.Chain_id.t -> Tezos_protocol_environment.rpc_context -> bytes -> bool -> callback_writer -> unit Directory.directory

val build_directory : string -> bool -> Tezos_mockup_registration.Registration.mockup_environment -> Tezos_base.TzPervasives.Chain_id.t -> Tezos_protocol_environment.rpc_context -> bytes -> unit Tezos_rpc.Directory.t

The directory of RPCs that the mockup client honors. Parameters are:

mem_only specifies whether the mockup uses a persistent state. mockup_env is the implementation provided by the protocol. chain_id is the only chain that the mockup honors. rpc_context is data used when honoring an RPC.