Parameter Make.E

include Tezos_mockup_registration.Registration.MOCKUP

type parameters

type protocol_constants

val parameters_encoding : parameters Tezos_base.TzPervasives.Data_encoding.t

val default_parameters : parameters

val protocol_constants_encoding : protocol_constants Tezos_base.TzPervasives.Data_encoding.t

val default_protocol_constants : Tezos_client_base.Client_context.full -> protocol_constants Tezos_base.TzPervasives.tzresult Lwt.t

The content equivalent to the default value of --protocol-constants

val default_bootstrap_accounts : Tezos_client_base.Client_context.full -> string Tezos_base.TzPervasives.tzresult Lwt.t

The content equivalent to the default value of --bootstrap-accounts

val protocol_hash : Tezos_base.TzPervasives.Protocol_hash.t

module Protocol : Tezos_mockup_registration.Registration_intf.PROTOCOL

module Block_services : sig ... end

val directory : Tezos_protocol_environment.rpc_context Tezos_rpc.Directory.t

val init : cctxt:Tezos_client_base.Client_context.printer -> parameters:parameters -> constants_overrides_json:Tezos_base.TzPervasives.Data_encoding.json option -> bootstrap_accounts_json:Tezos_base.TzPervasives.Data_encoding.json option -> Tezos_mockup_registration.Registration_intf.mockup_context Tezos_base.TzPervasives.tzresult Lwt.t

val migrate : Tezos_mockup_registration.Registration_intf.mockup_context -> Tezos_mockup_registration.Registration_intf.mockup_context Tezos_base.TzPervasives.tzresult Lwt.t

val chain_id : Tezos_base.TzPervasives.Chain_id.t

val rpc_context : Tezos_protocol_environment.rpc_context

val base_dir : string

val protocol_data : bytes