Module Mockup_args.Chain_id

val of_string : string -> Tezos_base.TzPervasives.Chain_id.t

of_string s converts any string to its valid chain_id representation. ** ** This function is useful to get valid, but dummy, chain identifiers out of ** command lines in the context of a mockup client.

val dummy : Tezos_base.TzPervasives.Chain_id.t

dummy is a valid dummy chain identifier. ** ** Do not use it outside of the context of a mockup client.

val choose : from_config_file:Tezos_base.TzPervasives.Chain_id.t option -> Tezos_base.TzPervasives.Chain_id.t

choose ~from_config_file ** produces a valid chain id from the optionally given one. ** ** The value from the command line, if any, has highest precedence, over the ** one from the config file. ** When both values are None, it uses the dummy.