Module Tezos_p2p.P2p_message

This module defines the messages of the P2p layers.

P2p messages are parameterized by a generic upper-layer message type 'msg. The P2p layer serializes these 'msg using a 'msg encoding' provided by the client of this lib.

The P2p protocol is simple and can be deduced mostly from the message type. To make the network topology more dynamic, it implements a simple peer swapping mechanism which works as follows.

Peer A sends a message : Swap_request (point, peer) to B. If B is already connected to the peer, the message is ignored. Otherwise B picks a peer peer' at point point' and connect to peer. If successful, it sends a response Swap_ack (point', peer') to A. Upon reception of Swap_ack. B tries to connected to peer'. If successful, it disconnect from peer.

type 'msg t =

| Bootstrap
(*
Welcome message sent by a peer upon connection
*)

| Advertise of Tezos_base.TzPervasives.P2p_point.Id.t list
(*
Response to a Bootstrap message, contains list of known points
*)

| Swap_request of Tezos_base.TzPervasives.P2p_point.Id.t * Tezos_base.TzPervasives.P2p_peer.Id.t
(*
Propose new peer/point and ask a peer/point to swap with
*)

| Swap_ack of Tezos_base.TzPervasives.P2p_point.Id.t * Tezos_base.TzPervasives.P2p_peer.Id.t
(*
Response to a swap request and propose peer/point to swap with.
*)

| Message of 'msg
(*
Generic upper-layer message
*)

| Disconnect
(*
Ending of connection, unused for now
*)

val encoding : 'a Tezos_base.TzPervasives.P2p_params.app_message_encoding list -> 'a t Tezos_base.TzPervasives.Data_encoding.t