Module Tezos_protocol_updater.Registered_protocol

module type T = sig ... end

type t = (module T)

val mem : Tezos_base.TzPervasives.Protocol_hash.t -> bool

val seq : unit -> t Tezos_base.TzPervasives.Seq.t

val get : Tezos_base.TzPervasives.Protocol_hash.t -> t option

get hash searches the protocol of the given hash in the registered protocols. This function also activates the logging facilty of the returned protocol.

val get_result : Tezos_base.TzPervasives.Protocol_hash.t -> t Tezos_base.TzPervasives.tzresult Lwt.t

Same as get but with the result return type.

val seq_embedded : unit -> Tezos_base.TzPervasives.Protocol_hash.t Tezos_base.TzPervasives.Seq.t

val get_embedded_sources : Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_base.TzPervasives.Protocol.t option

module Register_embedded_V0 (Env : Tezos_protocol_environment.V0.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V1 (Env : Tezos_protocol_environment.V1.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V2 (Env : Tezos_protocol_environment.V2.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V3 (Env : Tezos_protocol_environment.V3.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V4 (Env : Tezos_protocol_environment.V4.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V5 (Env : Tezos_protocol_environment.V5.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V6 (Env : Tezos_protocol_environment.V6.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V7 (Env : Tezos_protocol_environment.V7.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V8 (Env : Tezos_protocol_environment.V8.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V9 (Env : Tezos_protocol_environment.V9.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V10 (Env : Tezos_protocol_environment.V10.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V11 (Env : Tezos_protocol_environment.V11.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V12 (Env : Tezos_protocol_environment.V12.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V13 (Env : Tezos_protocol_environment.V13.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state

module Register_embedded_V14 (Env : Tezos_protocol_environment.V14.T) (Proto : Env.Updater.PROTOCOL) (Source : sig ... end) : T with type block_header_data = Proto.block_header_data and type operation_data = Proto.operation_data and type operation = Proto.operation and type operation_receipt = Proto.operation_receipt and type validation_state = Proto.validation_state