Module Env.Context_binary

type t = Tezos_context_memory.Context_binary.t

type tree = Tezos_context_memory.Context_binary.tree

module Tree : Context.TREE with type t = t with type tree = tree with type key = string list with type value = bytes

Tree provides immutable, in-memory partial mirror of the context, with lazy reads and delayed writes. The trees are Merkle trees that carry the same hash as the part of the context they mirror.

val make_empty_context : ?root:string -> unit -> t