Module Tezos_proxy.Registration

module type Proxy_sig = sig ... end

The module type of a proxy environment. Modules of this type should be prepared protocol-side and registered here to become available to the proxy facility.

type proxy_environment = (module Proxy_sig)

val register_proxy_context : proxy_environment -> unit

register_proxy_context env adds a proxy environment to the list of known environments. Raise Invalid_argument if an environment for the given protocol exists already.

val get_all_registered : unit -> proxy_environment list

get_all_registered () returns all proxy environments that have been registered so far with register_proxy_context

val get_registered_proxy : Tezos_client_base.Client_context.printer -> Tezos_base.TzPervasives.Protocol_hash.t -> proxy_environment Tezos_base.TzPervasives.tzresult Lwt.t

Returns a proxy environment for the given protocol.