Parameter Make.Request

type param

param represents the state/configuration of the service.

val initial_delay : Tezos_base.TzPervasives.Time.System.Span.t

initial_delay is a service configuration time constant. Typically, used to set up a retry time interval.

val active : param -> Tezos_base.TzPervasives.P2p_peer.Set.t

active param returns a set of active peers. For instance in order to broadcast a query to a set of peers.

val send : param -> Tezos_base.TzPervasives.P2p_peer.Id.t -> Hash.t list -> unit

send param peer_id kl queries peer with id peer_id for the values associated with the keys in kl.