Module type Requester.PROBE

When a requested value is received, it goes to a validation phase.

At fetching time, the client gives a param. At notification time, the client provides a notified_value. probe tries to construct a value from param and notified_value.

If no validation is needed. probe is defined as

let probe k () v -> Some v

type key

type param

type notified_value

type value

val probe : key -> param -> notified_value -> value option