Module Distributed_db.Operations

Index of all the operations of a given block (per validation pass).

For instance, fetch chain_db (block_hash, validation_pass) operation_list_list_hash queries the operation requester to get all the operations for block block_hash and validation pass validation_pass. It returns a list of operation, guaranteed to be valid with respect to operation_list_list_hash (root of merkle tree for this block).

type key = Tezos_base.TzPervasives.Block_hash.t * int

The index key

type value = Tezos_base.Operation.t list

The indexed data

val known : chain_db -> key -> bool Lwt.t

know t k returns true iff the key is present in the memory table or the disk.

type Tezos_base.TzPervasives.error +=

| Missing_data of key

type Tezos_base.TzPervasives.error +=

| Canceled of key

type Tezos_base.TzPervasives.error +=

| Timeout of key

val read : chain_db -> key -> value Tezos_base.TzPervasives.tzresult Lwt.t

Return value if it is found in-memory, or else on disk. Otherwise fail with error Missing_data.

val read_opt : chain_db -> key -> value option Lwt.t

Same as read but returns None if not found.

val inject : chain_db -> key -> value -> bool Lwt.t

inject t k v returns false if k is already present in the memory table or in the disk, or has already been requested. Otherwise it updates the memory table and return true

val fetch : chain_db -> ?peer:Tezos_base.TzPervasives.P2p_peer.Id.t -> ?timeout:Tezos_base.TzPervasives.Time.System.Span.t -> key -> Tezos_base.TzPervasives.Operation_list_list_hash.t -> value Tezos_base.TzPervasives.tzresult Lwt.t

fetch t ?peer ?timeout k param returns the value when it is known. It can fail with Timeout k if timeout is provided and the value isn't know before the timeout expires. It can fail with Cancel if the request is canceled.

The key is first looked up in memory, then on disk. If not present and not already requested, it schedules a request, and blocks until the requester is notified with notify. param is used to validate the notified value once it is received. (see also PROBE and notify).

Requests are re-sent via a 1.5 exponential back-off, with initial delay set to Request.initial_delay. If the function is called multiple time with the same key but with distinct peers, the internal scheduler randomly chooses the requested peer (at each retry).

val clear_or_cancel : chain_db -> key -> unit

Remove the data from the memory table if present. Otherwise decrements the number of watcher to this data. If their is only one watcher any pending fetch promises are resolved with the error Canceled.