Module Tezos_shell.Distributed_db_message

Tezos Shell - Network message for the gossip P2P protocol.

type t =

| Get_current_branch of Tezos_base.TzPervasives.Chain_id.t

| Current_branch of Tezos_base.TzPervasives.Chain_id.t * Tezos_base.Block_locator.t

| Deactivate of Tezos_base.TzPervasives.Chain_id.t

| Get_current_head of Tezos_base.TzPervasives.Chain_id.t

| Current_head of Tezos_base.TzPervasives.Chain_id.t * Tezos_base.Block_header.t * Tezos_base.Mempool.t

| Get_block_headers of Tezos_base.TzPervasives.Block_hash.t list

| Block_header of Tezos_base.Block_header.t

| Get_operations of Tezos_base.TzPervasives.Operation_hash.t list

| Operation of Tezos_base.Operation.t

| Get_protocols of Tezos_base.TzPervasives.Protocol_hash.t list

| Protocol of Tezos_base.Protocol.t

| Get_operations_for_blocks of (Tezos_base.TzPervasives.Block_hash.t * int) list

| Operations_for_block of Tezos_base.TzPervasives.Block_hash.t * int * Tezos_base.Operation.t list * Tezos_base.TzPervasives.Operation_list_list_hash.path

| Get_checkpoint of Tezos_base.TzPervasives.Chain_id.t

| Checkpoint of Tezos_base.TzPervasives.Chain_id.t * Tezos_base.Block_header.t

| Get_protocol_branch of Tezos_base.TzPervasives.Chain_id.t * int

| Protocol_branch of Tezos_base.TzPervasives.Chain_id.t * int * Tezos_base.Block_locator.t

| Get_predecessor_header of Tezos_base.TzPervasives.Block_hash.t * int32

| Predecessor_header of Tezos_base.TzPervasives.Block_hash.t * int32 * Tezos_base.Block_header.t

We must make sure that we are able to associate a response to the corresponding request. Thus, we are able to discrimate between expected responses from the unexpected ones.

val encoding : t Tezos_base.P2p_params.app_message_encoding list

val distributed_db_versions : Tezos_version.Distributed_db_version.t list

val cfg : Tezos_version.Distributed_db_version.Name.t -> t Tezos_base.P2p_params.message_config

val pp_json : Stdlib.Format.formatter -> t -> unit

module Bounded_encoding : sig ... end