Module Tezos_shell.Prevalidation

This module is used by the prevalidator worker to determine whether an operation should be added to the mempool. To do so, it calls on the protocol's Mempool to validate the operation, as well as the Prevalidator_bounding to ensure that the mempool remains bounded. It also retrieves relevant information from the protocol plugin.

module type T = sig ... end

module Make (Proto : Tezos_validation.Protocol_plugin.T) : T with type protocol_operation = Proto.operation and type chain_store = Tezos_store.Store.chain_store

How-to obtain an instance of this module's main module type: T