Module Proto.Plugin

type config

val config_encoding : config Tezos_base.TzPervasives.Data_encoding.t

val default_config : config

type info

Static internal information needed by pre_filter.

It depends on the head block upon which a mempool is built.

val init : Tezos_protocol_environment.Context.t -> head:Tezos_base.Block_header.shell_header -> info Tezos_base.TzPervasives.tzresult Lwt.t

Create an info based on the head block.

Should be called only once when a new prevalidator is started for a new protocol. Subsequent infos should be created using flush.

val flush : info -> head:Tezos_base.Block_header.shell_header -> info Tezos_base.TzPervasives.tzresult Lwt.t

Create a new info based on the head block.

Parts of the old info (which may have been built on a different block) are recycled, so that this function is more efficient than init and does not need a Tezos_protocol_environment.Context.t argument.

val syntactic_check : operation -> [ `Well_formed | `Ill_formed ] Lwt.t

Perform some syntactic checks on the operation.

To be used mostly as an exceptional mechanism to prevent ill-formed operations to block block application.

Note that the functions exposed in the output of proto_with_validation_plugin already call syntactic_check when appropriate.

val pre_filter : info -> config -> operation -> [ `Passed_prefilter of [ `High | `Medium | `Low of Q.t list ] | `Branch_delayed of Tezos_base.TzPervasives.tztrace | `Branch_refused of Tezos_base.TzPervasives.tztrace | `Refused of Tezos_base.TzPervasives.tztrace | `Outdated of Tezos_base.TzPervasives.tztrace ] Lwt.t

Perform some light preliminary checks on the operation.

If successful, return `Passed_prefilter with the priority of the operation, based on the operation kind and potentially its fee, gas, and size. If not, return a classification containing the encountered error.

Should be called on arrival of an operation and after a flush of the prevalidator.

val conflict_handler : config -> Mempool.conflict_handler

Return a conflict handler for Mempool.add_operation.

See the documentation of type Mempool.conflict_handler in e.g. lib_protocol_environment/sigs/v8/updater.mli.

Precondition: both operations must be individually valid (required by the protocol's operation comparison on which the implementation of this function relies).

module Conflict_map : sig ... end

The purpose of this module is to provide the fee_needed_to_replace_by_fee function. For this function to be correct, the caller must maintain the state of type t by calling update on each successfully validated operation and its induced replacements.

val fee_needed_to_overtake : op_to_overtake:operation -> candidate_op:operation -> int64 option

Compute the minimal fee (expressed in mutez) that candidate_op would need to have in order to be strictly greater than op_to_overtake according to compare_operations.

Return None when at least one operation is not a manager operation.

Also return None if both operations are manager operations but there was an error while computing the needed fee. However, note that this cannot happen when both manager operations have been successfully validated by the protocol.

type ctxt

Protocol context

val get_context : Tezos_protocol_environment.Context.t -> head:Tezos_base.TzPervasives.Block_header.shell_header -> ctxt Tezos_base.TzPervasives.tzresult Lwt.t

Return the protocol context

val sources_from_operation : ctxt -> operation -> Tezos_crypto.Signature.public_key_hash list Lwt.t

Return the sources from the operation