Module T.Internal_for_tests

val get_mempool_operations : t -> protocol_operation Tezos_base.TzPervasives.Operation_hash.Map.t

Return the map of operations currently present in the protocol representation of the mempool.

type mempool

Type Tezos_protocol_environment.PROTOCOL.Mempool.t.

val set_mempool : t -> mempool -> t

Modify the mempool field of the internal state t.

type bounding_state

Type Prevalidator_bounding.T.state.

val get_bounding_state : t -> bounding_state

Return the bounding_state component of the state.

val set_bounding_state : t -> bounding_state -> t

Modify the bounding_state component of the state.