Module Tezos_shell.Prevalidator_bounding

Bound the valid operations in the mempool by limiting both their cardinal and their total byte size.

type config = {

max_operations : int;
(*
Maximal allowed number of valid operations in the mempool.
*)

max_total_bytes : int;
(*
Maximal allowed sum of the byte sizes of all valid operations in the mempool.
*)

}

Mempool bounds.

They can be retrieved/set using RPCs GET/POST /chains/<chain>/mempool/filter.

val default_max_total_bytes : int

Default max_total_bytes is 10_000_000.

A block can have at most around 700k bytes of operations (see validation_passes in proto_xxx/lib_protocol/main.ml). So with this bound, a mempool can have the content of more than 10 blocks, which is more than enough.

val default_max_operations : int

Default max_operations is 10_000.

The smallest operations are around 140 bytes (e.g. 139 bytes for a (pre)attestation, 146 bytes for a delegation) so a block can have at most around 5_000 operations. But many operations are much larger, so in practice a block contains much less operations, so keeping 10_000 of them in the mempool is enough.

val default_config : config

Default bounds.

val config_encoding : config Tezos_base.TzPervasives.Data_encoding.t

Encoding for config.

It is internally an object without any variable field, so it is possible to use it in a Data_encoding.merge_objs.

module type T = sig ... end

Interface of a mempool bounding module.

module Make (Proto : Tezos_protocol_environment.PROTOCOL) : T with type protocol_operation = Proto.operation

Build a mempool bounding module.

module Internal_for_tests : sig ... end