Module Prevalidator_bounding.Internal_for_tests

module type T = sig ... end

Module type that includes T above, but is also aware of the exact definition of the state, so that the tests can access its fields and check its invariants.

module Make (Proto : Tezos_protocol_environment.PROTOCOL) : T with type protocol_operation = Proto.operation