Module type Shell_operation.PARSER

The purpose of this module type is to provide the parse function, whose return type depends on the protocol.

type protocol_operation

Similar to the same type in the protocol, see Tezos_protocol_environment.PROTOCOL.operation

val parse : Tezos_base.TzPervasives.Operation_hash.t -> Tezos_base.Operation.t -> protocol_operation operation Tezos_base.TzPervasives.tzresult

parse hash op reads a usual Operation.t and lifts it to the type protocol_operation used by this module. This function is in the tzresult monad, because it can return the following errors:

Validation_errors.Oversized_operation if the size of the operation data within op is too large (to protect against DoS attacks), and

Validation_errors.Parse_error if serialized data cannot be parsed.