Module Tezos_shell.Validator

Tezos Shell - Main entry point of the validation scheduler.

type t

val create : Tezos_store.Store.t -> Distributed_db.t -> Tezos_shell_services.Shell_limits.peer_validator_limits -> Tezos_shell_services.Shell_limits.block_validator_limits -> Block_validator_process.t -> Tezos_shell_services.Shell_limits.prevalidator_limits -> Tezos_shell_services.Shell_limits.chain_validator_limits -> start_testchain:bool -> t Tezos_base.TzPervasives.tzresult Lwt.t

val shutdown : t -> unit Lwt.t

val activate : t -> start_prevalidator:bool -> validator_process:Block_validator_process.t -> Tezos_store.Store.chain_store -> Chain_validator.t Tezos_base.TzPervasives.tzresult Lwt.t

Start the validation scheduler of a given chain.

val get : t -> Tezos_base.TzPervasives.Chain_id.t -> Chain_validator.t Tezos_base.TzPervasives.tzresult

val get_active_chains : t -> Tezos_base.TzPervasives.Chain_id.t list

val validate_block : t -> ?force:bool -> ?chain_id:Tezos_base.TzPervasives.Chain_id.t -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.Operation.t list list -> (Tezos_base.TzPervasives.Block_hash.t * unit Tezos_base.TzPervasives.tzresult Lwt.t) Tezos_base.TzPervasives.tzresult Lwt.t

Force the validation of a block.

val watcher : t -> Tezos_store.Store.Block.t Lwt_stream.t * Lwt_watcher.stopper

Monitor all the valid block (for all activate chains).

val chains_watcher : t -> (Tezos_base.TzPervasives.Chain_id.t * bool) Lwt_stream.t * Lwt_watcher.stopper

val inject_operation : t -> ?chain_id:Tezos_base.TzPervasives.Chain_id.t -> force:bool -> Tezos_base.Operation.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

inject_operation t ?chain_id ~force op notifies the prevalidator worker associated with the chain_id of a new injected operation. Fails if the branch in op is not known in the given chain_id or if no prevalidator is associated with the chain_id. If no chain_id is provided, try to recover a chain_id from the branch in the operation data. Fails if no chain_id can be recovered unless force is set. If force is set, notify all the known prevalidator workers.

val distributed_db : t -> Distributed_db.t