Module Tezos_shell_benchmarks.Encoding_benchmarks

module No_codegen : sig ... end

module Make_elliptic_curve_encoding_benchmarks (A : sig ... end) : sig ... end

module Ed25519 : sig ... end

module Secp256k1 : sig ... end

module P256 : sig ... end

module Bls : sig ... end

val chain_id_encoding : Tezos_benchmark.Benchmark.simple_with_num

val chain_id_decoding : Tezos_benchmark.Benchmark.simple_with_num

val chain_id_readable_encoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val chain_id_readable_decoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

The following benchmarks are not used directly to define precise gas models but the inferred gas constants may be used for over-approximating other costs or simply as reference points.

val purpose : Tezos_benchmark.Benchmark.purpose

val nat_encoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val nat_decoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val nat_decoding_intercept : (module Tezos_benchmark.Benchmark.Simple_with_num)

val int_encoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val int_decoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val int_decoding_intercept : (module Tezos_benchmark.Benchmark.Simple_with_num)

val string_encoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val string_decoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val adversarial_string_list_encoding : (module Tezos_benchmark.Benchmark.Simple_with_num)

val adversarial_string_list_decoding : (module Tezos_benchmark.Benchmark.Simple_with_num)