Module Tezos_shell_context

module Proxy_delegate_maker : sig ... end

module Shell_context : sig ... end

Shell_context is a persisting recursive map backed by an on-disk Irmin tree, i.e. Tezos_context.Context (in src/lib_context), used by nodes. It is the "main" kind of context. It abstracts away Tezos_context.Context, by hiding the internals, and making it an instance of the generic Protocol_environment.Context.t. Other *_context modules of Tezos_protocol_environment, i.e. siblings of this file, are backed by different type of values coming from Tezos_context.