Module Tezos_shell_services.Block_services

module Proof = Tezos_context_sigs.Context.Proof_types

type version =

| Version_1

| Version_2

type chain = [

| `Main

| `Test

| `Hash of Tezos_base.TzPervasives.Chain_id.t

]

type chain_prefix = unit * chain

val chain_path : (unit, chain_prefix) Tezos_rpc.Path.t

val parse_chain : string -> (chain, string) Stdlib.result

val chain_to_string : chain -> string

val chain_arg : chain Tezos_rpc.Arg.t

type block = [

| `Genesis
(*
The genesis block
*)

| `Head of int
(*
The nth predecessor of the current_head block if n > 0. If n = 0, represents the current_head. n should not be negative since the current_head does not have successors.
*)

| `Alias of [ `Caboose | `Checkpoint | `Savepoint ] * int
(*
The nth predecessor of the caboose, the checkpoint or the savepoint if n > 0. If n = 0, represents the block itself. If n < 0, represents the nth successor.
*)

| `Hash of Tezos_base.TzPervasives.Block_hash.t * int
(*
The nth predecessor of the block of given hash if n > 0. If n = 0, represents the block itself. Otherwise, if n < 0, represents the nth successor.
*)

| `Level of Stdlib.Int32.t
(*
The block at a given level
*)

]

A representation of a block's position relatively to a known block of a chain.

val parse_block : string -> (block, string) Stdlib.result

type range = [ `Level of Stdlib.Int32.t ] * [ `Level of Stdlib.Int32.t ]

val parse_block_range : string -> (range, string) Stdlib.result

A block range in the form level..level. Currently, this function supports only level..level.

type block_or_range =

| Block of block

| Range of range

val parse_block_or_range : string -> (block_or_range, string) Stdlib.result

val to_string : block -> string

type prefix = (unit * chain) * block

val dir_path : (chain_prefix, chain_prefix) Tezos_rpc.Path.t

val path : (chain_prefix, chain_prefix * block) Tezos_rpc.Path.t

val mempool_path : ('a, 'b) Tezos_rpc.Path.t -> ('a, 'b) Tezos_rpc.Path.t

val live_blocks_path : ('a, 'b) Tezos_rpc.Path.t -> ('a, 'b) Tezos_rpc.Path.t

type operation_list_quota = {

max_size : int;

max_op : int option;

}

val raw_context_encoding : Proof.raw_context Tezos_base.TzPervasives.Data_encoding.t

val raw_context_insert : (string list * Proof.raw_context) -> Proof.raw_context -> Proof.raw_context

raw_context_insert (k,v) c inserts a key-value pair (k,v) in a raw_context c. If k collides to a existing sub-tree in c, the sub-tree is replaced by a new key-value pair.

module type PROTO = sig ... end

type protocols = {

current_protocol : Tezos_base.TzPervasives.Protocol_hash.t;

next_protocol : Tezos_base.TzPervasives.Protocol_hash.t;

}

val protocols : Tezos_rpc.Context.simple -> ?chain:chain -> ?block:block -> unit -> protocols Tezos_base.TzPervasives.tzresult Lwt.t

module Make (Proto : PROTO) (Next_proto : PROTO) : sig ... end

module Fake_protocol : PROTO

module Empty : sig ... end