Module Tezos_shell_services.Injection_services

type Tezos_base.TzPervasives.Error_monad.error +=

| Injection_operations_error

While injecting several operations (see private_operations), if one injection failed, we have to report the error. To avoid using recursive errors we do the reporting as follows:

We wrap the error into Injection_operations_error

If injecting the operation oph succeeded we use Injection_operation_succeed_case oph

If injecting the operation oph failed with err, we use Injection_operation_error_case oph followed by err.

type Tezos_base.TzPervasives.Error_monad.error +=

| Injection_operation_succeed_case of Tezos_base.TzPervasives.Operation_hash.t

type Tezos_base.TzPervasives.Error_monad.error +=

| Injection_operation_error_case of Tezos_base.TzPervasives.Operation_hash.t

val block : Tezos_rpc.Context.simple -> ?async:bool -> ?force:bool -> ?chain:Chain_services.chain -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.Operation.t list list -> Tezos_base.TzPervasives.Block_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

block cctxt ?async ?force raw_block tries to inject raw_block inside the node. If ?async is true, raw_block will be validated before the result is returned. If ?force is true, the block will be injected even on non strictly increasing fitness.

val operation : Tezos_rpc.Context.simple -> ?async:bool -> ?chain:Chain_services.chain -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Operation_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

val private_operation : Tezos_rpc.Context.simple -> ?async:bool -> ?chain:Chain_services.chain -> Tezos_base.TzPervasives.Bytes.t -> Tezos_base.TzPervasives.Operation_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

val private_operations : Tezos_rpc.Context.simple -> ?async:bool -> ?force:bool -> ?chain:Chain_services.chain -> Tezos_base.TzPervasives.Bytes.t list -> Tezos_base.TzPervasives.Operation_hash.t list Tezos_base.TzPervasives.tzresult Lwt.t

private_operations injects multiple operations. The private_ prefix is because the service is bound to the /private/ path-prefix which is intended for tests only. See the ~description argument in the definition in the ml file for more information.

val protocol : Tezos_rpc.Context.simple -> ?async:bool -> Tezos_base.Protocol.t -> Tezos_base.TzPervasives.Protocol_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

module S : sig ... end