Module Tezos_shell_services.Preapply_result

type 'error t = {

applied : (Tezos_base.TzPervasives.Operation_hash.t * Tezos_base.Operation.t) list;
(*
Applied operations
*)

refused : (Tezos_base.Operation.t * 'error list) Tezos_base.TzPervasives.Operation_hash.Map.t;
(*
Refused operations, for example because of an invalid signature
*)

outdated : (Tezos_base.Operation.t * 'error list) Tezos_base.TzPervasives.Operation_hash.Map.t;
(*
Outdated operations, for example a late attestation
*)

branch_refused : (Tezos_base.Operation.t * 'error list) Tezos_base.TzPervasives.Operation_hash.Map.t;
(*
Branch refused operations, for example because of insufficient balance
*)

branch_delayed : (Tezos_base.Operation.t * 'error list) Tezos_base.TzPervasives.Operation_hash.Map.t;
(*
Branch delayed operations, for example because of a timestamp in the future
*)

}

A container for classified operations

val empty : 'error t

empty t returns an empty instance of t: all collections * are initialized empty

val encoding : 'error list Tezos_base.TzPervasives.Data_encoding.t -> 'error t Tezos_base.TzPervasives.Data_encoding.t

encoding err_enc t, where err_enc is the encoding for errors, * returns an encoding for t