Module Prevalidator_worker_state.Operation_encountered

Indicates how an operation hash has been encountered:

Injected : The corresponding operation has been directly injected into the node.

Notified peer_id : The hash is present in a mempool advertised by peer_id.

Arrived : The node was fetching and has just received the corresponding operation.

Other serves as default value for an argument, but in practice it is not used atm (June 2021).

This module is used in Event.Banned_operation_encountered.

type situation =

| Injected

| Arrived

| Notified of Tezos_base.P2p_peer_id.t option

| Other

type t = situation * Tezos_base.TzPervasives.Operation_hash.t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val pp : Stdlib.Format.formatter -> t -> unit