Module Prevalidator_worker_state.Request

type ('a, 'b) t =

| Flush : Tezos_base.TzPervasives.Block_hash.t * Chain_validator_worker_state.update * Tezos_base.TzPervasives.Block_hash.Set.t * Tezos_base.TzPervasives.Operation_hash.Set.t -> (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
The chain changed, the mempool is being notified of the new state.
*)

| Notify : Tezos_base.P2p_peer.Id.t * Tezos_base.Mempool.t -> (unit, Tezos_base.TzPervasives.Empty.t) t
(*
The given peer sent this mempool.
*)

| Leftover : (unit, Tezos_base.TzPervasives.Empty.t) t
(*
Operations not yet processed should be processed.
*)

| Inject : {

op : Tezos_base.Operation.t;

force : bool;

} -> (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
Operation has been locally injected (sent) to the node.
*)

| Arrived : Tezos_base.TzPervasives.Operation_hash.t * Tezos_base.Operation.t -> (unit, Tezos_base.TzPervasives.Empty.t) t
(*
Operation was fetched by the node.
*)

| Advertise : (unit, Tezos_base.TzPervasives.Empty.t) t
(*
Current mempool should be advertised to all known peers.
*)

| Ban : Tezos_base.TzPervasives.Operation_hash.t -> (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t
(*
User requested the node to ban operation with this hash.
*)

type view =

| View : (_, _) t -> view

val view : (_, _) t -> view

val encoding : view Tezos_base.TzPervasives.Data_encoding.t

val pp : Stdlib.Format.formatter -> view -> unit