Module Tezos_shell_services.Shell_limits

The configurable constants used by shell components as maximum, with their encodings and default values.

type operation_metadata_size_limit =

| Unlimited

| Limited of int

States whether or not the operation metadata size must be caped and potentially discard if the given size limit is exceeded.

val operation_metadata_size_limit_encoding : operation_metadata_size_limit Tezos_base.TzPervasives.Data_encoding.t

type block_validator_limits = {

protocol_timeout : Tezos_base.Time.System.Span.t;

operation_metadata_size_limit : operation_metadata_size_limit;

}

val default_block_validator_limits : block_validator_limits

val block_validator_limits_encoding : block_validator_limits Tezos_base.TzPervasives.Data_encoding.t

type prevalidator_limits = {

max_refused_operations : int;
(*
The maximum number of operations tracked by the mempool for each of the refused, branch delayed, branch refused and outdated operation classifications. Default is 1000
*)

operation_timeout : Tezos_base.Time.System.Span.t;
(*
The maximum time allowed to fetch the contents of an operation advertised by a remote peer. Default is 10 seconds
*)

operations_batch_size : int;
(*
Maximum number of pending operations processed (or classified) at the end of each request to the prevalidator worker. Default is 50
*)

}

This record contains the different limits and settings that can be updated from a node configuration for a prevalidator

val default_prevalidator_limits : prevalidator_limits

Sane default values for limits

val prevalidator_limits_encoding : prevalidator_limits Tezos_base.TzPervasives.Data_encoding.t

type peer_validator_limits = {

new_head_request_timeout : Tezos_base.Time.System.Span.t;

block_header_timeout : Tezos_base.Time.System.Span.t;

block_operations_timeout : Tezos_base.Time.System.Span.t;

protocol_timeout : Tezos_base.Time.System.Span.t;

}

val default_peer_validator_limits : peer_validator_limits

val peer_validator_limits_encoding : peer_validator_limits Tezos_base.TzPervasives.Data_encoding.t

type synchronisation_limits = {

latency : int;
(*
latency is the time interval (seconds) used to determine if a node is synchronized with a chain. For instance, a node that knows head with timestamp T is synchronized if T >= now - max_latency. This parameter depends on the baking rate and the latency of the network.
*)

threshold : int;
(*
threshold determines the number of peers the synchronization heuristic looks at to determine if the node is synchronized or not.
*)

}

Constants parameterizing the bootstrap heuristics.

type chain_validator_limits = {

synchronisation : synchronisation_limits;

}

val default_chain_validator_limits : chain_validator_limits

val chain_validator_limits_encoding : chain_validator_limits Tezos_base.TzPervasives.Data_encoding.t

val default_disable_context_pruning : bool

val default_storage_maintenance_delay : Storage_maintenance.delay

type limits = {

block_validator_limits : block_validator_limits;

prevalidator_limits : prevalidator_limits;

peer_validator_limits : peer_validator_limits;

chain_validator_limits : chain_validator_limits;

history_mode : History_mode.t option;

disable_context_pruning : bool option;

storage_maintenance_delay : Storage_maintenance.delay option;

}

val default_limits : limits

val limits_encoding : limits Tezos_base.TzPervasives.Data_encoding.t