Module Store.Utilities

Utilities brings some utility functions that aims to help extracting values, data or statistics from the store.

val stat_metadata_cycles : t -> (string * Tezos_store_shared.Store_types.metadata_stat list) list Tezos_base.TzPervasives.tzresult Lwt.t

Store utility function that aims to give an overview of the shape of the store's metadata.