Module Tezos_store_shared.Block_repr

Block representation effectively stored on disk and its accessors.

Type definitions and encodings

type contents = {

header : Tezos_base.TzPervasives.Block_header.t;

operations : Tezos_base.TzPervasives.Operation.t list list;

block_metadata_hash : Tezos_base.TzPervasives.Block_metadata_hash.t option;

operations_metadata_hashes : Tezos_base.TzPervasives.Operation_metadata_hash.t list list option;

}

The type for the effective contents of a block is its header and the operations it contains. Their metadata hashes are also present.

type metadata = {

message : string option;

max_operations_ttl : int;

last_preserved_block_level : Stdlib.Int32.t;

block_metadata : Tezos_base.TzPervasives.Bytes.t;

operations_metadata : Tezos_validation.Block_validation.operation_metadata list list;

}

The type for a block's metadata stored on disk. This representation is tightly linked to Tezos_validation.Block_validation.result which also has a strong dependency to Tezos_protocol_environment.validation_result.

Some fields exposed by Tezos_validation.Block_validation.result are unnecessary hence the lack of direct link.

type block = {

hash : Tezos_base.TzPervasives.Block_hash.t;

contents : contents;

mutable metadata : metadata option;

}

The type for a block stored on disk.

The hash of the block is also stored to improve efficiency by not forcing the user to hash the header. This also allows to store fake hashes (e.g. sandbox's genesis blocks) but should be prevented by the API.

The metadata might not be present. The mutability flag allows users to re-use the same structure to store freshly loaded metadata.

type t = block

Genesis

val create_genesis_block : genesis:Tezos_base.TzPervasives.Genesis.t -> Tezos_base.TzPervasives.Context_hash.t -> t

create_genesis_block ~genesis context_hash creates a default genesis block for the given genesis and its context_hash that contains metadata.

val contents_encoding : contents Tezos_base.TzPervasives.Data_encoding.t

Encoding for contents.

val metadata_encoding : metadata Tezos_base.TzPervasives.Data_encoding.t

Encoding for metadata.

val equal : t -> t -> bool

Equality on block

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

Encoding for t (and block).

Important An encoded block is prefixed by 4 bytes which stands for the length of the data. This is the case with Data_encoding.dynamic_size ~kind:`Uint30 encodings. This will be expected to be present to improve the store efficiency.

val pp_json : Stdlib.Format.formatter -> t -> unit

pp_json pretty-print a block as JSON.

Accessors

val descriptor : t -> Store_types.block_descriptor

descriptor block returns the pair (hash x level) of block.

val hash : t -> Tezos_base.TzPervasives.Block_hash.t

hash block returns the stored block's hash. It is not guaranteed to be the same as Block_header.hash (header block) (e.g. in sandbox, the genesis block might have a fake hash).

val operations : t -> Tezos_base.TzPervasives.Operation.t list list

operations block returns the list of list of operations contained in the block.

Block header accessors

val header : t -> Tezos_base.TzPervasives.Block_header.t

val shell_header : t -> Tezos_base.TzPervasives.Block_header.shell_header

val level : t -> Stdlib.Int32.t

val proto_level : t -> int

val predecessor : t -> Tezos_base.TzPervasives.Block_hash.t

val timestamp : t -> Tezos_base.TzPervasives.Time.Protocol.t

val validation_passes : t -> int

val operations_hash : t -> Tezos_base.TzPervasives.Operation_list_list_hash.t

val fitness : t -> Tezos_base.TzPervasives.Fitness.t

val context : t -> Tezos_base.TzPervasives.Context_hash.t

val protocol_data : t -> Tezos_base.TzPervasives.Bytes.t

val block_metadata_hash : t -> Tezos_base.TzPervasives.Block_metadata_hash.t option

val operations_metadata_hashes : t -> Tezos_base.TzPervasives.Operation_metadata_hash.t list list option

Metadata accessors

val metadata : t -> metadata option

val message : metadata -> string option

val max_operations_ttl : metadata -> int

val last_preserved_block_level : metadata -> Stdlib.Int32.t

val block_metadata : metadata -> bytes

val operations_metadata : metadata -> Tezos_validation.Block_validation.operation_metadata list list

Utility functions

val check_block_consistency : ?genesis_hash:Tezos_base.TzPervasives.Block_hash.t -> ?pred_block:t -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

check_block_consistency ?genesis_hash ?pred_block block checks that the stored data is consistent:

Does the hash stored equals the result of Block_header.hash of its header and, if not, is this the stored genesis_hash?

Is the block a successor of pred_block with regards to its level and its predecessor's hash?

Are the stored operations hashes consistent regarding the stored operations hashes?

val decode_metadata : string -> metadata option

decode_metadata data decodes metadata from data encoded either with the new encoding or the legacy one.