Module Tezos_store_shared.Store_types

Global types used in the store library

module Block_store_status : sig ... end

type block_descriptor = Tezos_base.TzPervasives.Block_hash.t * int32

The type used to describe a block pointer i.e. its hash and level.

val block_descriptor_equal : block_descriptor -> block_descriptor -> bool

Equality function for block_descriptor.

val block_descriptor_encoding : block_descriptor Tezos_base.TzPervasives.Data_encoding.t

Encoding for block_descriptor.

type chain_config = {

history_mode : Tezos_shell_services.History_mode.t;

genesis : Tezos_base.TzPervasives.Genesis.t;

expiration : Tezos_base.TzPervasives.Time.Protocol.t option;

}

The type for chain configuration.

val chain_config_equal : chain_config -> chain_config -> bool

Equality function for chain_config.

val chain_config_encoding : chain_config Tezos_base.TzPervasives.Data_encoding.t

Encoding for chain_config.

val pp_block_descriptor : Stdlib.Format.formatter -> block_descriptor -> unit

Pretty-printer for block_descriptor

type invalid_block = {

level : int32;

errors : Tezos_base.TzPervasives.Error_monad.error list;

}

The type used to store an invalid block's value. We only retain the level and the errors encountered during validation. These values should be indexed by the block's hash.

val invalid_block_equal : invalid_block -> invalid_block -> bool

Equality function on invalid_block. Warning: uses polymorphic comparison on errors.

val invalid_block_encoding : invalid_block Tezos_base.TzPervasives.Data_encoding.t

Encoding for invalid_block.

module Block_lru_cache : Aches_lwt.Lache.MAP_OPTION with type key = Tezos_base.TzPervasives.Block_hash.t

Module Block_lru_cache implements a lwt LRU cache map indexed by block hashes.

module Protocol_levels : sig ... end

Module Protocol_levels represents an associative map of protocol levels to corresponding blocks which supposedly activate new protocols, that is to say blocks that acknowledge a protocol change in the next block.

type metadata_stat = {

block_level : Stdlib.Int32.t;

block_metadata_size : Stdlib.Int64.t;

operation_metadata_arity : int list;

operation_metadata_size : Stdlib.Int64.t;

manager_operation_metadata_sizes : int list;

too_large_operation_metadata_count : Stdlib.Int64.t;

}