Module Tezos_store_unix.Consistency

val check_consistency : [ `Chain_dir ] Tezos_store_shared.Naming.directory -> Tezos_base.TzPervasives.Genesis.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

check_consistency store_dir genesis checks that the store is in a consistent state.

val fix_consistency : ?history_mode:Tezos_shell_services.History_mode.t -> [ `Chain_dir ] Tezos_store_shared.Naming.directory -> Tezos_context_ops.Context_ops.index -> Tezos_base.TzPervasives.Genesis.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

fix_consistency ?history_mode store_dir context_index genesis aims to fix a store if it is in an inconsistent state. To do so, it will inspect the stored data and try infer all the missing or corrupted parts.