Module Floating_block_index.Block_info

Module Block_info represents the values stored in the index which are used to locate the position of a block in a floating block store and its predecessors.

val max_predecessors : int

The maximum number of predecessors stored per block.

type t = {

offset : int;
(*
offset in the file
*)

predecessors : Tezos_base.TzPervasives.Block_hash.t list;
(*
predecessors of the block
*)

resulting_context_hash : Tezos_base.TzPervasives.Context_hash.t;
(*
context hash after block application
*)

}

The type for storing the block's info.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-printer for t