Module Tezos_store_unix.Protocol_store

Protocol store

type t

The type for the protocol store.

val mem : t -> Tezos_base.TzPervasives.Protocol_hash.t -> bool

mem pstore proto_hash tests the existence of the protocol indexed by proto_hash in the store.

val all : t -> Tezos_base.TzPervasives.Protocol_hash.Set.t

all pstore returns the set of all stored protocols in pstore.

val raw_store : t -> Tezos_base.TzPervasives.Protocol_hash.t -> bytes -> Tezos_base.TzPervasives.Protocol_hash.t option Lwt.t

raw_store pstore proto_hash proto_bytes stores on disk the protocol proto_bytes (encoded bytes) indexed as proto_hash. Returns None if the protocol already exists.

val store : t -> Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_base.TzPervasives.Protocol.t -> Tezos_base.TzPervasives.Protocol_hash.t option Lwt.t

store pstore proto_hash protocol stores on disk the protocol protocol indexed as proto_hash. Returns None if the protocol already exists.

val read : t -> Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_base.TzPervasives.Protocol.t option Lwt.t

read pstore proto_hash reads from pstore and returns the protocol indexed by proto_hash. Returns None if the protocol cannot be read.

val init : [ `Store_dir ] Tezos_store_shared.Naming.directory -> t Lwt.t

init store_dir creates a store relatively to store_dir path or loads it if it already exists.