Module Tezos_validation

module Block_validation : sig ... end

This module is a proxy for the shell of the protocol (for the application part). The main function of this module is apply whichs calls the one of the protocol.

module Command_line : sig ... end

module External_validation : sig ... end

module External_validator : sig ... end

module Patch_context : sig ... end

module Protocol_plugin : sig ... end