Module type Protocol_plugin.HTTP_CACHE_HEADERS

Protocol specific plugin to expose helper functions in the implementation of Http cache headers RPC middleware.

val hash : Tezos_base.TzPervasives.Protocol_hash.t

val get_round_end_time : get_context:(unit -> Tezos_protocol_environment.Context.t Lwt.t) -> Tezos_base.Block_header.shell_header -> Tezos_base.TzPervasives.Time.System.t option Lwt.t

get_round_end_time ctx curr_header gets the time at which the current round ends which is the time at which the next round starts. Useful to get an estimate of when the next block should arrive.