Module type Protocol_plugin.METRICS

This is a protocol specific module that is used to collect all the * protocol-specific metrics. This module * allows to decode protocol data payload and provide back basic * types that can be used as metrics.

val hash : Tezos_base.TzPervasives.Protocol_hash.t

val update_metrics : protocol_metadata:bytes -> Tezos_base.TzPervasives.Fitness.t -> (cycle:float -> consumed_gas:float -> round:float -> unit) -> unit Lwt.t