Module Tezos_mockup.Files

Mockup file abstraction

type t = private string

File path

val get_mockup_directory : dirname:string -> t

mockup_directory ~dirname gets the full name of the mockup subdirectory located inside dirname

val exists_mockup_directory : dirname:string -> bool Lwt.t

has_mockup_directory ~dirname checks whether a mockup subdirectory exists in directory dirname

module type ACCESSOR = sig ... end

Generic signature to access files located inside mockup subdirectory

module Context : ACCESSOR

Accessor for context.json

module Mempool : ACCESSOR

Accessor for mempool.json

module Trashpool : ACCESSOR

Accessor for thrashpool.json